requirements.txt
uvicorn[standard]
fastapi
jpype1
xmltodict
lru-dict